Added extended precision and complex number support packages.

+++ DecimalNumber/decNumber.c
@@ -352,7 +352,7 @@ decNumber * decNumberFromUInt32(decNumber *dn, uInt uin) {
     *up=(Unit)(uin%(DECDPUNMAX+1));
     uin=uin/(DECDPUNMAX+1);
     }
-  dn->digits=decGetDigits(dn->lsu, up-dn->lsu);
+  dn->digits=decGetDigits(dn->lsu, (Int)(up-dn->lsu));
   return dn;
   } // decNumberFromUInt32
 
@@ -654,7 +654,7 @@ decNumber * decNumberFromString(decNumber *dn, const char chars[],
     #if DECDPUN>1
     out=0;                         // accumulator
     up=res+D2U(d)-1;               // -> msu
-    cut=d-(up-res)*DECDPUN;        // digits in top unit
+    cut=(Int)(d-(up-res)*DECDPUN);        // digits in top unit
     for (c=cfirst;; c++) {         // along the digits
       if (*c=='.') continue;       // ignore '.' [don't decrement cut]
       out=X10(out)+(Int)*c-(Int)'0';
@@ -832,7 +832,7 @@ decNumber * decNumberAnd(decNumber *res, const decNumber *lhs,
       } // both OK
     } // each unit
   // [here uc-1 is the msu of the result]
-  res->digits=decGetDigits(res->lsu, uc-res->lsu);
+  res->digits=decGetDigits(res->lsu, (Int)(uc-res->lsu));
   res->exponent=0;                      // integer
   res->bits=0;                          // sign=0
   return res;  // [no status to set]
@@ -1219,7 +1219,7 @@ decNumber * decNumberInvert(decNumber *res, const decNumber *rhs,
       } // each digit
     } // each unit
   // [here uc-1 is the msu of the result]
-  res->digits=decGetDigits(res->lsu, uc-res->lsu);
+  res->digits=decGetDigits(res->lsu, (Int)(uc-res->lsu));
   res->exponent=0;                      // integer
   res->bits=0;                          // sign=0
   return res;  // [no status to set]
@@ -1845,7 +1845,7 @@ decNumber * decNumberOr(decNumber *res, const decNumber *lhs,
       } // non-zero
     } // each unit
   // [here uc-1 is the msu of the result]
-  res->digits=decGetDigits(res->lsu, uc-res->lsu);
+  res->digits=decGetDigits(res->lsu, (Int)(uc-res->lsu));
   res->exponent=0;                      // integer
   res->bits=0;                          // sign=0
   return res;  // [no status to set]
@@ -2551,7 +2551,7 @@ decNumber * decNumberRotate(decNumber *res, const decNumber *lhs,
           } // whole units to rotate
         // the rotation may have left an undetermined number of zeros
         // on the left, so true length needs to be calculated
-        res->digits=decGetDigits(res->lsu, msumax-res->lsu+1);
+        res->digits=decGetDigits(res->lsu, (Int)(msumax-res->lsu+1));
         } // rotate needed
       } // rhs OK
     } // numerics
@@ -3277,7 +3277,7 @@ decNumber * decNumberXor(decNumber *res, const decNumber *lhs,
       } // non-zero
     } // each unit
   // [here uc-1 is the msu of the result]
-  res->digits=decGetDigits(res->lsu, uc-res->lsu);
+  res->digits=decGetDigits(res->lsu, (Int)(uc-res->lsu));
   res->exponent=0;                      // integer
   res->bits=0;                          // sign=0
   return res;  // [no status to set]
@@ -6541,11 +6541,11 @@ static Int decUnitAddSub(const Unit *a, Int alength,
 
   // OK, all A and B processed; might still have carry or borrow
   // return number of Units in the result, negated if a borrow
-  if (carry==0) return c-clsu;     // no carry, so no more to do
+  if (carry==0) return (Int)(c-clsu);     // no carry, so no more to do
   if (carry>0) {                   // positive carry
     *c=(Unit)carry;                // place as new unit
     c++;                           // ..
-    return c-clsu;
+    return (Int)(c-clsu);
     }
   // -ve carry: it's a borrow; complement needed
   add=1;                           // temporary carry...
@@ -6568,7 +6568,7 @@ static Int decUnitAddSub(const Unit *a, Int alength,
     *c=(Unit)(add-carry-1);
     c++;                      // interesting, include it
     }
-  return clsu-c;              // -ve result indicates borrowed
+  return (Int)(clsu-c);              // -ve result indicates borrowed
   } // decUnitAddSub
 
 /* ------------------------------------------------------------------ */
@@ -6752,7 +6752,7 @@ static Int decShiftToLeast(Unit *uar, Int units, Int shift) {
   if (cut==DECDPUN) {              // unit-boundary case; easy
     up=uar+D2U(shift);
     for (; up<uar+units; target++, up++) *target=*up;
-    return target-uar;
+    return (Int)(target-uar);
     }
 
   // messier
@@ -6780,7 +6780,7 @@ static Int decShiftToLeast(Unit *uar, Int units, Int shift) {
     count-=cut;
     if (count<=0) break;
     }
-  return target-uar+1;
+  return (Int)(target-uar+1);
   } // decShiftToLeast
 
 #if DECSUBSET
@@ -7643,7 +7643,7 @@ static decNumber *decDecap(decNumber *dn, Int drop) {
   cut=MSUDIGITS(dn->digits-drop);       // digits to be in use in msu
   if (cut!=DECDPUN) *msu%=powers[cut];  // clear left digits
   // that may have left leading zero digits, so do a proper count...
-  dn->digits=decGetDigits(dn->lsu, msu-dn->lsu+1);
+  dn->digits=decGetDigits(dn->lsu, (Int)(msu-dn->lsu+1));
   return dn;
   } // decDecap
